Refactor FolioClient.__init__() and FolioClient.login() to improve testability
Instantiating the FolioAuth class directly in the __init__ method of FolioClient requires users of FolioClient to mock the entire class when writing tests that need an instance of FolioClient. If we shift the creation of the FolioAuth object into the login() method, users can just patch login() and create a mock FolioAuth object with the appropriate properties for their testing context.